Backend 1.4.0, Frontend 1.2.0
Backend has conceptual classes of Cards, decks, actions, players,
environment, the game "table", the order of player-turns, encapsulated
decks and trashpiles, serverside decision-making on card-draws.
Added example card-classes for market and village.
Deck-spec-protocol '1' builds an initial deck of 6 cards, of which 3 are
market and 3 are village.

Frontend has a dynamic view on the cards in hand, with a scrollbar that
only shows when needed.
NOTE: THE CARDS DISPLAYED ARE HARDCODED.