Frontend 1.2.2 Backend 1.4.2
Frontend changes:
> engine sends "playable" cards to the client, client makes those cards
glow, and binds an eventhandler to the card that sends the card
information to the server.
> server accepts cards from the player on turn. If the card that is
played is valid, every client gets a notification about the played card,
and the player who played it recevies a package that makes the client
remove the played card from his hand.